Revert "Result grade mark conversion use ToDouble instead"
This reverts commit a9354f373b61fdee78de83f57806812852794ae1.

+++ BUEnrolment/Models/Result.cs
@@ -31,19 +31,19 @@ namespace BUEnrolment.Models
                 NameValueCollection appSettings = ConfigurationManager.AppSettings;
                 ResultGrade grade;
 
-                if (Mark >= Convert.ToDouble(appSettings["HighDistinction"]))
+                if (Mark >= Convert.ToInt32(appSettings["HighDistinction"]))
                 {
                     grade = ResultGrade.HighDistinction;
                 }
-                else if (Mark >= Convert.ToDouble(appSettings["Distinction"]))
+                else if (Mark >= Convert.ToInt32(appSettings["Distinction"]))
                 {
                     grade = ResultGrade.Distinction;
                 }
-                else if (Mark >= Convert.ToDouble(appSettings["Credit"]))
+                else if (Mark >= Convert.ToInt32(appSettings["Credit"]))
                 {
                     grade = ResultGrade.Credit;
                 }
-                else if (Mark >= Convert.ToDouble(appSettings["Pass"]))
+                else if (Mark >= Convert.ToInt32(appSettings["Pass"]))
                 {
                     grade = ResultGrade.Pass;
                 }
